# OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Tutorial Management API",
    "description": "This API exposes endpoints to manage tutorials.",
    "termsOfService": "https://www.bezkoder.com/terms",
    "contact": {
      "name": "My Name",
      "url": "https://www.fct.unl.pt",
      "email": "myemail@gmail.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://choosealicense.com/licenses/mit/"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/",
      "description": "Server URL in Development environment"
    },
    {
      "url": "https://api.example.com/",
      "description": "Server URL in Production environment"
    }
  ],
  "paths": {
...
    "/reviews/{username}/review/{bookingId}": {
      "post": {
        "tags": [
          "review-controller"
        ],
        "operationId": "createReview",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Review was successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewInfoDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Review data.",
            "content": {}
          },
          "403": {
            "description": "Wrong client.",
            "content": {}
          },
          "404": {
            "description": "Booking not found.",
            "content": {}
          },
          "409": {
            "description": "Booking already reviewed.",
            "content": {}
          }
        }
      }
    },
...
    "/reviews/{reviewId}": {
      "get": {
        "tags": [
          "review-controller"
        ],
        "operationId": "getReview",
        "parameters": [
          {
            "name": "reviewId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review with given Id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewInfoDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden",
            "content": {}
          },
          "404": {
            "description": "Review not found.",
            "content": {}
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/reviews/": {
      "get": {
        "tags": [
          "review-controller"
        ],
        "operationId": "getAllReviews",
        "responses": {
          "200": {
            "description": "Found all reviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewInfoDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden",
            "content": {}
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
...
    "/clients/{username}/reviews": {
      "get": {
        "tags": [
          "client-controller"
        ],
        "operationId": "getClientReviews",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of client reviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewInfoDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Client not found.",
            "content": {}
          }
        }
      }
    },
...
    "/bookings/{bookingID}/review": {
      "get": {
        "tags": [
          "booking-controller"
        ],
        "operationId": "getBookingReview",
        "parameters": [
          {
            "name": "bookingID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Found booking review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewInfoDTO"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not reviewed yet.",
            "content": {}
          }
        }
      }
    },
...
    "/apartments/{apartmentId}/reviews": {
      "get": {
        "tags": [
          "apartment-controller"
        ],
        "operationId": "getApartmentReviews",
        "parameters": [
          {
            "name": "apartmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of Apartment's Reviews.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewInfoDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Apartment not found."
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateReviewDTO": {
        "required": [
          "description",
          "rating"
        ],
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ReviewInfoDTO": {
        "required": [
          "clientName"
        ],
        "type": "object",
        "properties": {
          "clientName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  }
}